+++ src/controllers/commands/builtin/twitch/Poll.hpp
@@ -13,4 +13,10 @@ namespace chatterino::commands {
 /// /poll
 QString createPoll(const CommandContext &ctx);
 
+/// /endpoll
+QString endPoll(const CommandContext &ctx);
+
+/// /cancelpoll
+QString cancelPoll(const CommandContext &ctx);
+
 }  // namespace chatterino::commands